requirements.txt
python-binance
python-dotenv
urllib3